pyproject.toml
[tool.poetry]
name = "lh3_operator_assignment"
version = "0.1.0"
description = ""
authors = ["guins2000 <salfjsdl@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
lh3api = "^0.2.0"
pandas = "^1.4.2"
lxml = "^4.8.0"
openpyxl = "^3.0.9"
dateparser = "^1.1.1"
XlsxWriter = "^3.0.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
